Enable setUnifiedTitleAndToolBarOnMac for all Qt versions.
We changed our minds, and this will return in (possibly) 5.2.1. For the
versions in which it is not available, it is simply a no-op and
therefore safe to use in any version.

When QtMacExtras has a more native solution available (which will not
be backported to Qt 4) this can go in the #else of an ifdef for
QT_MACEXTRAS_LIB.